Fail on non-200 statusCodes, and only forward on 3xx statuses.
This is increasingly relevant, because github seems to be throwing a lot of 500 errors today.  What I'd previously thought was oddness in the tar/childProc stuff in fact seems to be a result of downloading HTML error messages as tgz files, and then trying to unzip them, which of course failed.