Save ALL dependencies to a "dependson" folder
This is different than the "dependencies" folder.  That folder is for linking in modules.  However, there may be a case where a package exposes no main or lib, but is nevertheless required by some other package.  (Perhaps it uses its CLI executable or something.)

Also, having a link to the dependency's package dir is SUPER handy, as it makes uninstallation much simpler, which is very helpful in fixing #92