Relative links! Fix #104. Fix #21.
All shims and linking uses the relativize function to calculate the shortest relative path from the source to destination.  In cases where there is no common ancestor in the paths, or if the relative path is more characters than the absolute path, then the absolute path is used instead.

This should make npm a LOT easier to integrate with programs like ivy and nave, which keep the binaries and modules in a managed location.