Seems like everything's pretty much working now with the prototypal configs
This is a huge commit.  Sorry.  It's a massive refactor, and a lot of things changed.  Quite frankly, I'm feeling a bit too lazy to go through and meticulously tag each change with the thinking behind it.

The code isn't terribly hard to follow.  This is something that should've been done a long time ago.