Ugly kludge: Create the relevant folders as soon as we know what they are.
This is not a beautiful solution.  The configuration loader should not be concerned with folder creation.  For instance, this means that it might have to have write access to the npm root (to create the .tmp folder) even when just doing "npm ls" (which doesn't use the temp folder anyway.)

However, there is no other single place where npm accesses those folders at the moment, so this is the easiest way to ensure that these things exist.

TODO: Abstract out *all* the file reading/writing stuff to go through some single gate that ensures that the relevant folders exist at the point where they are actually needed.