Upon closer examination, fuck it.
If we've got shims being created for the lib folder, why not?  I mean, it's a pretty sensible default behavior.  The only thing that sucked about it was that it didn't work as expected, but now it does, so why not keep it?

If you *don't* want your lib directory to be exposed, just do `"modules":{}` or something.