Add the "logfd" config
Log statements are written to the file descriptor or stream object passed in as the "logfd" configuration param.

For bonus points, if it's not a tty, it doesn't get colorised.

So, if someone does this:

npm install npm --logfd 1 | pbcopy

then it'll actually be a log that can be pasted into an email without a bunch of tty color code noise.