etags, etags, etags!!
Cache all GET requests to the registry, and (at least in the default
registry.get entry point) investigate that cache file and do conditional
requests to the registry.

Furthermore, if a tarball is in cache, and we look for it again, and the
etag matches, then assume that the tarball is still valid, and don't
download it again.

The other place this is extra winful is on generating the views for ls.
That's getting pretty slow, and is much faster using the conditional
requests.