Mostly rewrite npm ls.  Completion wonderfulment!
Make tab-completion for npm ls way more awesome.  It's actually like a
dynamic search now, because it'll use the previous items in the list as
arguments to limit how many things it can match against the current arg.

As a pleasant side effect of this, it required making npm ls much more
programmatically useful, because now you can filter the returned data
based on a set of args in script just like how you can on the command
line, and get the same set of data back.