Fix #417.  chdir() doesn't parallelize well.
Removed all the process.chdir() stuff that was a relic from a time when
npm didn't do multiple different builds at the same time.

Not sure how this bug stayed alive for so long.  Maybe since it usually
works if you just try again, and it's hard to deterministically
reproduce.

Special thanks to Charlie Robbins for bringing it to my attention.