Add spermies, 2.x.x versions, and negative tests.
X-versions:

x.x.x matches anything
N.x.x matches from N.0.0 up to and not including (N+1).0.0.
N.M.x matches from N.M.0 up to and not including N.(M+1).0

If any section is empty, it'll be treated like an x.  So,

1 --> 1.x.x
1.x --> 1.x.x
1.2 --> 1.2.x

Spermies:

The spermy operator is "~>" or "~".  It does "sane" matching from the
specified version up through the version that increases the Major or
minor version number by 1.

This has the odd and slightly inelegant effect that ~1.2 and ~1.2.0 are
the same exact thing.  But, whatever.  It's a standard.  Blame rubygems.

~, ~>, ~>x, ~>x.x, ~x.x.x --> * (matches anything)
~1, ~1.x, ~1.x.x --> >=1.0.0 <2.0.0
~1.0, ~1.0.x --> >=1.0.0 <1.1.0
~1.2, ~1.2.x --> >=1.2.0 <1.3.0
~1.2.0 --> >=1.2.0 <1.3.0
~1.2.3 --> >=1.2.3 <1.3.0

In the end, just like the "1.2.3 - 2.4.5" version ranges, it's
transformed into a set of simple >,<,>=,<= comparators.

Also: Add a bunch of negative tests to validate that versions that
shouldn't satisfy a given range actually don't.  Long overdue on that
one.