Add cache completion for cache (ls|clean)
I forgot that I'd written this handy little find util.  I don't for the
life of me remember what it was for, and got halfway through writing
another one when I decided that it belonged in a separate file.

I told vim to ":e lib/utils/find.js", only to be surprised to find
the very utility that I'd meant to write.

It's a bit like finding a $20 in the pocket of a jacket that has been
ignored all summer.