Closes GH-691 Resolve linked deps when building tarballs
This makes it so that if you use `npm link` to install something locally
as a bundle, it'll resolve the folder and then snapshot it into your
local bundle, so that you publish what you're using right at the moment
of publishing.

Then, after doing that monkey business for the tarball creation, it puts
everything back the way it was.