Reverse-chronological order; consistent (dis)use of periods within each change; rewrote markup for improved readability.
…

HISTORY: 
Put history in reverse-chronological order (as the most common use case for this file is humans who want to see the most recent changes).  

PERIODS: 
Consistency is your friend (and mine!).

MARKUP: 
Each version is a subheading, and each change is a bullet point.  Although the previous method wasn't semantically inaccurate, this change fosters improved semantics as well as improved readability for humans.