Several readability-focused changes to consistency, markup, and formatting.
Example codeblocks consistently prefixed with **Good:** or **Bad:**. This makes it obvious at-a-glance whether the example code is intended to be emulated or avoided.  (Similarly, prefixed **Do:** and **Don't:** to guidelines under "Whitespace".)
Marked inline code with backticks.  (See the "Semicolons" section for example.)
Other various minor edits, all aiming for clarity of phrasing/formatting.