Only use tilde range descriptor when package >=1.0
Problem: When using `install --save`, npm prefixed the installed
package version with the tilde range descriptor. This makes sense
for packages that follow semver and have reached 1.0.0. But for
packages < 1.0.0 semver allows APIs to change freely, so those
should be packaged with the exact version that was installed,
as implemented by this patch.