Don't prematurely exit, ever
I think this is why the npm cache gets corrupted so often.
If there's a failure while another package is in the process
of downloading, then it'll result in a corrupted tarball.

Instead, just set the exitCode, and only actually exit with
tht code when the 'exit' event occurs, so that in-transit
jobs are not forcibly killed.