Replace the log util with npmlog module
This feels so good.  lib/utils/log.js is the worst kind of
glue code that keeps the npm project from being properly
abstracted into independent pieces.

In the process, also cleaned up a lot of unproductive logging,
and made the npm-debug.log generated on errors be a bit more
easy to read.