Revert "Quote paths passed to child_process.exec."
This reverts commit 391b8a3239710efaa9089dab8d5e2b2f25e19b6b.

This causes problems on unix when running the tests, which is much more
common.

The test runner needs to probably just be redone in a better way.  It's
old and janky, and has been ported from bash to js, but ought to be
replaced with something that exercises npm's functionality in a more
proper way.