Fix #2538 Properly ignore false ancestors
When running `npm install` in a package, if one of the deps depends on
the root package, it might not be require()-able (if the folder name
doens't match, and/or if it's not in a node_modules folder).

This makes that work.

Originally by @SubStack, edited slightly by @isaacs.

* failing test for pkg.name that can't be required
* fix the false_name test by altering data.name when it's not in a
  requirable location
* set data.name to null for unrequirable packages, less hackishly