Make version and name optional (mostly)
Fix #3146

Version and name are still required to add to the cache folder,
or to publish, so you'll have to name your packages before installing
them in some other place.

However, this cleans up the process a bit for master branches and folks
using branches for releases that don't get folded back in.