Separate tar and cache lockfile usage
Folders were getting double-locked, resulting in an unnecessary 1000ms
hang for everything added to cache.  Since we tar things multiple times
(stupidly) this led to about a 5-10s increase in latency for lots of
operations.

Fix that by adding 'tar://' to the start of all the locks created in
lib/utils/tar.js