graceful-fs@1.2.3
Related to #3259

Because readFile and writeFile were not wrapped, there were cases where
a lot of package.json reading would cause problems.  This wasn't an
issue with earlier versions of graceful-fs, because it monkey-patched
the builtin FS module's open() method, so that readFile and writeFile
would Just Work.

However, that caused other problems.

Note that fs.ReadStream and fs.WriteStream are still not covered in the
open fd guard.  That can be done later, but this extends EMFILE coverage
to the more commonly used methods first.