the tests require docs to be made already
Otherwise the `make test` command fails to find ronn when generating the
docs for the first time in the prepublish step.

Probably there's a better way to do this, since tarball generation
requires a dev dep to be installed.  But this works on Unix at least,
and gets it by for now.