do not add root pkg to ancestors
The ancestors list is only relevant for detecting pathological cycles in
the dep history.  However, the root package is often NOT actually
require()-able, so is not in the family list, but will (reasonably) be
in the dependency hash of a package being installed.  So, omit it as an
ancestor.

Fixes installing `lru-cache` in development mode.  No non-dev-mode
use-cases have been encountered.