Allow npm bugs for the current directory
If you are in your local module, you might want to visit the repository without
knowing the repo name or typing the name in.

This adds `npm bugs .` and `npm bugs` so that the npm-bugs-command behaves like
`npm install`, `npm docs`, `npm home`, and `npm repo.`

Also fixed the parsing of `./` (instead of `.`).

Closes #4204.