Fix tests
the tests were depending on a warm cache, as npm was calling to
a server on 1331 but the mock-registry was serving packages from
another port. now a temporary cache location is used to avoid
that the tests don't fail as they are getting everything from
the cache.

the server which served the metadata was removed, the metadata is
now served by the mock-registry itself, so npm can also install
from there.

furthermore the package metadata was separated in files.