adduser: allow change of the saved password
Modify `npm adduser` to allow the user to change the password used
for authentication. This is useful when the password was changed
via the website and the user wants to update local npm config
to use this new password.

Before this change, `npm adduser` would not ask for a password when
the username was not changed, thus it was not possible to easily
change the password while keeping the same username.