Allow to build all the docs OOTB.
Since 1.4.10, the html docs are generated via the marked module. when
doing the usual "make" dance, the build process stops when attempting to
generate the first html doc complaining about a missing
node_modules/.bin/marked.

This patch is a quick'n'dirty copy-pasta of the treatment applied to
load the ronn module but applied to marked. When applied against the
1.4.13 tarball it allows -- at least in a works-for-me manner -- to
generate all the docs.