Only chmodr git remotes in Windows
Recursively setting file permissions on git remotes breaks git
hooks for anyone who uses git templates and symlinks their hooks
across repositories by removing execute permissions on the shared
hooks.

Since the original commit that added this feature:
https://github.com/npm/npm/commit/48263bf152c4e4842961cedb75d381b1d6fe3985
was a fix for a Windows only issue:
https://github.com/npm/npm/issues/3117

The proposed solution provided in this PR is to only apply the
file permissions in Windows.