Don't squash execute perms in _git-remotes/ dir
A previous fix for https://github.com/npm/npm/issues/3117
explicitly set the permissions for all files to a hard-coded
value, regardless of what the permissions were previously.

This changes that behavior to *add* the permissions that were
previously being *set* so that existing execute permissions are
maintained.

This is a redo of the much more blunt solution attempted in
https://github.com/npm/npm/pull/5728
thanks to some very helpful feedback from @othiym23