Print original version range instead of expanded one on incompatible-version error
This address Problem 2 of #6180. On an incompatible-version error, instead of
printing the expanded range of the user's desired range (`>=1.9.0 <1.10.0` as
opposed to `~1.9`), `npm` will now print the original range so that there's no
confusion on the user's part as to what happened to their input.

So, instead of this:

    $ npm install phantomjs@~1.9
    npm ERR! notarget No compatible version found: phantomjs@'>=1.9.0 <1.10.0'

The user will see this:

    $ npm install phantomjs@~1.9
    npm ERR! notarget No compatible version found: phantomjs@'~1.9'

Closes #6199