rm -rf npm submodule command
Noticed today that 'npm submodule' is broken in npm 2.0.

Went digging, and thought 'Hey, how did this EVER work?'

Dug a bit deeper, and found that this command was actually let go way
back in the summer of 2013, and had been hanging around ever since due
to a glitch in the system.

So, with this patch, we've 'fixed the glitch', removing it from the set
of npm commands.

Fix #6369