inflight installs, prevent peerDependency problems
Because of the cheesy backdoor mechanics peerDependencies use to get
themselves installed, it's possible that more than one install to the
same place can get queued up when peerDependencies aren't otherwise
depended on. See the comment in the commit for details, and start
getting used to explicitly depending on your peerDependencies, because
that's the future. The future is saner than now.