never remove project directories (fixes #5754)
A failed install could cause the removal of an empty (new) project
directory because fs-vacuum wasn't being told where to stop vacuuming up
empty directories. Pass in the project path to provide a hard
backstop for fs-vacuum, and be more careful to specify which things
should be gently removed during unbuilding.