badly-named man files won't install (fixes #7000)
Before, if you tried to put a filename in the `man` stanza of
package.json that didn't follow the rules as listed in `npm help 5
package.json`, installing would work, but upgrading or uninstalling
would fail with a cryptic failure. Now, fail on build with a descriptive
message, and log the same descriptive message as an error (but don't
fail) on unbuild.