scripts: use full path for cmd.exe on Win32
Currently lifecycle.js assumes that Window's cmd.exe is on the PATH,
and fails with a spawn ENOENT error if it is not.

The Windows 'comspec' environment variable contains the full filepath
to the default command interpreter, eg "C:\Windows\System32\cmd.exe".
Should it not be set, we fall-back to using 'cmd' from PATH, as before.

Fixes https://github.com/npm/npm/issues/5267