Fix spelling error in README
Line 176 uses the word "warrantee" where the word "warranty" should be used instead. 

According to http://grammarist.com/, "A warrantee is someone to whom a warranty is given. A warranty is a guarantee given by a company to a purchaser stating that a product will meet certain requirements during a given time."

It may seem like I am nitpicking, but I believe this is worth attention considering this word misuse is occurring in the legal section of the README, which can potentially open this open source project up to legal loop holes in the future is someone were so inclined.