don't use $HOME/.npm as cache directory when running as root
When installing packages globally using root we can't use $HOME/.npm as
$HOME is most often not changed by sudo and so it's still the user's own
home dir.  Now, if a local npm install is run it may fail since folders in
the cache dir are owned by root. In particular when the local npm install
tries to grab newer package versions. Instead we install to the temp
dir when run as root.