test: eliminate fixture directories from test/tap
It's too difficult for npm contributors to figure out what the
conventional style is for tests. Part of the problem is that the
documentation in CONTRIBUTING.md is inadequate, but another important
factor is that the tests themselves are written in a variety of styles.
One of the most notable examples of this is the fact that many tests use
fixture directories to store precooked test scenarios and package.json
files.

This had some negative consequences:

  * tests weren't idempotent
  * subtle dependencies between tests existed
  * new tests get written in this deprecated style because it's not
    obvious that the style is out of favor
  * it's hard to figure out why a lot of those directories existed,
    because they served a variety of purposes, so it was difficult to
    tell when it was safe to remove them

All in all, the fixture directories were a major source of technical
debt, and cleaning them up, while time-consuming, makes the whole test
suite much more approachable, and makes it more likely that new tests
written by outside contributors will follow a conventional style. To
support that, all of the tests touched by this changed were cleaned up
to pass the `standard` style checker.