Improve shasum check sanity
Previously, npm's behavior in case it did not receive a shasum was to:

1. Itself generate a shasum for whatever package is downloaded over HTTP,
2. Cache this arbitrary shasum as a valid, genuine shasum, exactly as if it was sent via TLS from `registry.npmjs.com`.

I would argue that this behavior is not reasonable. Instead, npm should insist on a shasum being delivered and do diligent comparison for any requested packages, without exception.

I am a new contributor to this project, so please excuse me in case I am missing something. I have done some testing, however, and feel this solution is legitimate. I also look forward to investigating in the future the possibility of adding feasible, easy to use code-signing features into npm.