version: add lifecycle scripts around the version command
Fixes #7906. This was requested to allow authors to require scripts to be
true prior to bumping the version number for release. For instance, by
running the test suite.

As bumping the version number is a typical action immediately prior to
releasing a new version, this acts as a `prepublish` that ties to the actual
act of publication, versus the `prepublish` lifecycle, which is run under
some install-time circumstances as well.