bug: Fix #8138 stack-overflow when idealizing large trees
What was going on here is that we were shortcircuiting when
there weren't any children... and these calls ended up essentially
chaining together one after another till the call stack limit
as hit.  Insisting that this function be async w/ dezalgo allows
the short cirtuiting to not blow the stack.