install: Prefer rawSpec over spec when saving
This is necessary because npa mangles spec in ways that are not
compatible npa itself. For reasons. Specifically, `git+ssh` urls have
their `git+` stripped from the front, which, for reasons, is then
fed BACK into npa, which then explodes.

Relatedly, we can't just use rawSpec because it's not available when someone
doesn't include a version at all, where as spec will be '*'.

PR-URL: https://github.com/npm/npm/pull/9174
Fixes: #9077