doc: remove misleading API documentation
The API is awkward, brittle, and frequently counterintuitive. Having it
documented sets an unrealistic expectation of it being a supported /
encouraged thing for developers outside the team to use. Until we have
an API we should stand behind, having it be undocumented is actually a
net win because people won't be misled into thinking they're gonna have
a good time by using it.