test: Fix the progress-config test when run from travis or other CI
We have special code to disable the progress bar on travis (and other CI) to
save ourselves and our users a lot of unhelpful output. Unfortunately as
this test didn't take that into account, it immediately exploded in this
envs.

So we fix that up, PLUS we add tests for the various env options for
suppressing progress bars.

PR-URL: https://github.com/npm/npm/pull/9304