lifecycle: keep private values out of child procs
npm@1.x didn't include configuration values prefixed with an underscore
in the environment passed to chiled processes. npm@2.x includes the
notion of scoped configuration, where a nerfed URL can be used to prefix
configuration, some of which might be underscore-prefixed without the
scope. Add that behavior to npm@2 and close this regression.

PR-URL: https://github.com/npm/npm/pull/9348