Document aliases for npm cache
'npm cache clear' does not fail like `npm cache boom` does, but it
doesn't provide any output either. What does it do?

A StackOverflow answer suggests it's not the same, but the source code
shows it's a silent alias for `npm cache clean`, along with `npm cache
rf`.

Documenting these aliases clears up of the confusion.

The aliases for 'ls' are now documented as well, and the formatting of
the page is improved slightly as well.