logical-tree: Make it mutate to improve performance
The clone we were doing to save ourselves from mutation proved to be
excessively slow. While lodash is being updated to not have the
same scaling issues, it was determined that we weren't gaining
anything from not mutating here, so there was no reason to pay
even a lesser price.