deps: fix interactions between modules managed by shrinkwraps and those not
Shrinkwraps need to let you override dependency versions specified in
package.json files. Indeed, this was already supported, but it was a bit
to keen on this.  Previously, if you had a dep in your shrinkwrap then anything
that required that dependency would count as a match, regardless of version.

We're changing this so that it only does this broad matching when the
matched module is a direct child of the module doing the requiring.

PR-URL: https://github.com/npm/npm/pull/10153