deps: Make failure of deps of extraneous deps fail
When a non-optional dep fails, we have to walk up the requiredBy tree to see
if its descended only from optional dependencies. If it ISN'T then
we fail at this point, if it IS then we just mark those ancestors and
their children as failed and call it a day.

This brings us to the case of modules the user installed but didn't --save.
These modules don't have any ancestors from a requiredBy perspective and as
such, we never hit the top level module which is where we ordinarily stop.
As such, we need to short circuit– if a descendent of a user installed dep
fails then we need to fail the install.

Credit: @iarna
PR-URL: https://github.com/npm/npm/pull/10465
Reviewed By: @othiym23