bundles: Refactor the bundled module model (again)
This rejiggers bundled modules, such that they're moved out of the bundling
module's folder at extract time and places in staging by themselves like any
other module. This allows the rest of npm to pretend they're normal modules,
which lets us remove a whole host of special cases and overall simplifies
our code.

Fixes: #10482

So if you had:
```
a bundles b requres c
```
And the version of bundled `c` in `b` doesn't match the `package.json` it
will end up installing a new copy of `c` under `b` and ditching the contents
of `b`.