add-remote-tarball: work around node 0.8 http back-pressure bug
0.8 http streams have a bug, where if they're paused with data in
their buffers when the socket closes, they call `end` before emptying
those buffers, which results in the entire pipeline ending and thus
the point that applied backpressure never being able to trigger a
`resume`.

We work around this by piping into a pass through stream that has
unlimited buffering. The pass through stream is from readable-stream
and is thus a current streams3 implementation that is free of these
bugs even on 0.8.

PR-URL: https://github.com/npm/npm/pull/10903
Credit: @iarna