Change working directory for `npm run scriptname` in pkg.json
This commit adds the above mentioned feature by checking whether the package
JSON includes a scriüt in the form of...
{
  "scripts": {
    "change-cwd-test": {
       "cwd": "test/fixtures/",
       "run": "node -e \"console.log(process.cwd())\""
    }
   }
}
It then attaches a un-enumerable array to the `scripts` objects, which the `npm`
client can check for during execution.
From the `npm` client's POV, this seems a little odd since it needs to transform
a lot. But in order to pass hard checks here and w/ `npm` and not to break, this
seemed to be necessary for the author.
*Background:*
this in greater detail. Fundamentally though, this for enabling `npm` to be more
of a build tool. For example some programs such as make don't allow non-cwd
runs. And in order to not `cd` to it, this is supposed to greatly help.

Credit: @eljefedelrodeodeljefe
Reviewed-By:
PR-URL: npm#