shrinkwrap: understand and exclude optionalDependencies
Optional dependencies tend to be dependencies that are nice to include,
but aren't required for the operation of a project. Commonly used for
dependencies that are needed only on one platform.

If a shrinkwrap is created on a platform where an optional dependency is
installed, the optional dependencies are then included in the
shrinkwrap. These dependencies may fail to install on other platforms,
causing the whole install to be in an indeterminate state.

This patch adds logic to detect optional dependencies while turning the
dependency tree into a shrinkwrap file as well as excludes them from the
shrinkwrap.