cache: measure cache age from process start time
The default cache-min time is 10 seconds, which is shorter than the
average install time. This means that if a package has multiple
dependencies that depend on the same package there is a good chance that
a default install will result in the sub-dependencies being fetched
multiple times instead of using the copy that was put in to the cache
earlier _in the same install_.

Extend the cache-min to be at least the current age of the node process
so that objects fetched more than cache-min seconds ago are not
considered stale unless they are also older than the process.

This should save a few seconds off any moderately complex package
installation.

Testing with a fake package with an obscene number of tiny duplicate
sub-dependencies, the savings are small but noticeable even with a 100mb
connection:
 - before: 7m6.014s
 - after: 6m38.291s